Refactor methods extension
