+++ Crosscutting.Base/ApplicationBuilderExtensions.cs
@@ -1,0 +1,4 @@
+﻿namespace Crosscutting.Base;
+public static class ApplicationBuilderExtensions
+{
+}
