+++ Crosscutting.Base/Configurations/SerilogSettingsBinder.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Crosscutting.Base.Configurations
+﻿namespace Crosscutting.Base.Configurations
 {
     public class SerilogSettingsBinder
     {
